fix storj-sim env issue

+++ tests.sh
@@ -1,8 +1,15 @@
 #!/usr/bin/env bash
 set -ueo pipefail
 
+usage="usage: $0 <serialized API key>"
+
+if [[ $# -ne 1 ]]; then
+    echo ${usage}
+    exit 1
+fi
+
 export TMPDIR=/tmp/
 export SATELLITE_0_ADDR=127.0.0.1:10000
-export GATEWAY_0_API_KEY=$(storj-sim network env | grep GATEWAY_0_API_KEY | sed 's,GATEWAY_0_API_KEY=,,')
+export GATEWAY_0_API_KEY=$1
 
 make tests && ./tests && rm ./tests
